schedulerdb 5-of-5: Builders, everything else
This contains the rest of the main schedulerdb changes. Most of the changes
are in buildbot/process/builder.py, which asks the database for unclaimed
BuildRequests instead of maintaining a .buildable attribute: this allows
queued build requests to survive a master bounce. Once a build is started,
the Builder must claim the BuildRequests by marking them in the database, and
builds that complete are similarly updated. Much of the IBuilderControl
functionality changed, since now the database is used to manipulate queued
builds.

buildbot/master.py also changed extensively. The database manager object is
created in the constructor and made available to several other services.
Builders and Schedulers are now Service instances, so they must be handled
differently during config reloads. A separate SchedulerManager class is the
parent for Schedulers, and manages the Loop which drives them. The Deferred
that loadConfig() returns *must* be waited upon: the actual config changes
will not occur right away, so many unit tests need to be updated. Some of the
scheduling functionality of the BotMaster was moved, as were things like
submitBuildSet. Builds are always driven with BuildSets now (previously they
were optional), so the IControl methods are different, requiring updates in
other places.

The Status implementation was changed slightly, since BuildSets,
SourceStamps, and Changes now live in the database instead of pickles. Some
methods were added to retrieve DB-specific information (like unique ID
numbers) for these objects, or to exchange an ID number for row contents. The
top-level Status object uses DB notification callbacks to implement existing
subscription interfaces.